0.0.5
Исправлена команда получения тх и рх
Исправлена команда отправки тр и рх
Исправлен цикл фор на костантный

+++ messageCreatrors/statemessagesender.h
@@ -28,7 +28,7 @@ public:
 private:
     quint16 CalculateInt(double fvcoFreq) const;
     quint32 CalculateFract(double fvcoFreq) const;
-    quint8 CalculateGain(quint8 gain) const;
+    quint8 CalculateGain(double gain) const;
     quint8 CalculateAttenuator(quint16 attenuator) const;
     bool CalculateDiv(double fvcoFreq) const;
 private:
